rename of backend classes; add optional doxygen generation for future refactoring
